+++ dep/Libft/arr_handle/ft_strlen.c
@@ -22,5 +22,5 @@ size_t	ft_strlen(const char *str)
 	{
 		i++;
 	}
-	return (i);
+	return ((size_t)i);
 }
