+++ dep/Libft/arr_handle/ft_strmapi.c
@@ -23,7 +23,7 @@ char	*ft_strmapi(const char *s, char (*f)(unsigned int, char))
 		return (NULL);
 	while (s[i] != '\0')
 	{
-		ret[i] = f(i, s[i]);
+		ret[i] = f((unsigned int)i, s[i]);
 		i++;
 	}
 	ret[i] = '\0';
